third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/CustomElementRegistry-initialize-expected.txt

This is a testharness.js-based test.
[FAIL] initialize sets element.customElementRegistry to the global registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] initialize does not set the registry of nested shadow tree to the global registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] initialize sets element.customElementRegistry to a scoped registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] initialize does not set the registry of nested shadow tree to a scoped registry
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] initialize sets element.customElementRegistry permantently
  assert_equals: expected null but got object "[object CustomElementRegistry]"
[FAIL] initialize works on Document
  Failed to execute 'initialize' on 'CustomElementRegistry': The provided value is not of type '(Element or ShadowRoot)'.
[FAIL] initialize works on DocumentFragment
  Failed to execute 'initialize' on 'CustomElementRegistry': The provided value is not of type '(Element or ShadowRoot)'.
Harness: the test ran to completion.
